% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityCounts-package.R
\docType{package}
\name{activityCounts-package}
\alias{activityCounts}
\alias{activityCounts-package}
\title{activityCounts: Generating ActiLife Counts}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Creates ActiLife counts based on the raw acceleration data.
}
\keyword{internal}
