% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{counts}
\usage{
counts(data, hertz)
}
\arguments{
\item{data}{input accelerometer data, x y z}

\item{hertz}{sampling frequency, Hz}
}
\value{
Returns a \code{data.table} with three columns: \describe{
\item{x}{the number of counts for X axis}
\item{y}{the number of counts for Y axis}
\item{z}{the number of counts for Z axis}
}
}
\description{
Calculated ActiLife counts based on raw accelerometer data
}
\examples{
{
\donttest{counts(sampleXYZ,100)}
}

}
\seealso{
\link{sampleXYZ}
 raw accelerometer data for testing \code{counts()} function.

 \link{sampleCounts}
 counts calculated by a\code{ctivityCounts} and ActiLife
}
