% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptrunc.R
\name{pptrunc}
\alias{pptrunc}
\title{pptrunc}
\usage{
pptrunc(data, max_value)
}
\arguments{
\item{data}{The variable that will be truncated}

\item{max_value}{The upper bound ( -max_value is the lower bound)}
}
\value{
the highest(or the lowest) value of "data" and "max_value"
}
\description{
pptrunc
}
