% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsum.R
\name{runsum}
\alias{runsum}
\title{runsum}
\usage{
runsum(data, len, threshold)
}
\arguments{
\item{data}{input data}

\item{len}{the length}

\item{threshold}{the threshold}
}
\value{
returns a
}
\description{
runsum
}
