% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCounts.R
\docType{data}
\name{sampleCounts}
\alias{sampleCounts}
\title{The counts calculated by activityCounts and ActiLife based on included raw accelerometer data}
\format{A \code{data.table} with nine columns: \describe{
 \item{Raw_Accelerometer_X}{Raw accelerometer data in X direction}
 \item{Raw_Accelerometer_Y}{Raw accelerometer data in Y direction}
 \item{Raw_Accelerometer_Z}{Raw accelerometer data in Z direction}
 \item{activityCounts_x_counts}{counts calculated by \code{counts()} function in X direction}
 \item{activityCounts_y_counts}{counts calculated by \code{counts()} function in Y direction}
 \item{activityCounts_z_counts}{counts calculated by \code{counts()} function in Z direction}
 \item{ActiLife_x_counts}{counts calculated by ActiLife software in X direction}
 \item{ActiLife_y_counts}{counts calculated by ActiLife software in Y direction}
 \item{ActiLife_z_counts}{counts calculated by ActiLife software in Z direction}
 }}
\usage{
sampleCounts
}
\description{
A simple \code{data.table} which its three first rows are raw accelerometer data for testing the
 \code{\link{counts}} function. The first three rows are identical to the other included dataset, \code{sampleXYZ},  Sampling frequency is 100H. The second three rows are counts calculated by the
  package \code{activityCounts} and the last three row are counts calculated by the ActiLife software.
}
\seealso{
\link{counts}
 to see how to produce counts.

 \link{sampleXYZ}
 raw accelerometer data for testing \code{counts()} function.
}
\keyword{datasets}
