% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc.R
\name{trunc}
\alias{trunc}
\title{trunc}
\usage{
trunc(data, min_value)
}
\arguments{
\item{data}{The input variable which will be altered if less than the threshold}

\item{min_value}{the threshold which the input below it will be set to zero}
}
\value{
returns zero if the "data" is less than the "mean_value" otherwise returns the "data"
}
\description{
trunc
}
