% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityCounts-package.R
\docType{package}
\name{activityCounts-package}
\alias{activityCounts}
\alias{activityCounts-package}
\title{activityCounts: Generating ActiLife Counts}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Creates ActiLife counts based on the raw acceleration data.
}
\details{
There are two datasets and a main function in this package
}
\seealso{
\link{counts}
 to see how to produce counts.

 \link{sampleXYZ}
 raw accelerometer data for testing \code{counts()} function.

 \link{activityCounts}
 raw accelerometer data for testing \code{counts()} function.
}
\author{
Ruben Brondeel, Javad Rahimipour Anaraki, SeyedJavad KhataeiPour, Daniel Fuller.

\subsection{Maintainers}{SeyedJavad KhataeiPour <skhataeipour@mun.ca>,
 Javad Rahimipour Anaraki <jra066@mun.ca>, Daniel Fuller <dfuller@mun.ca>
 }
}
\concept{The main function anf sample datasets}
\keyword{internal}
