% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityCounts-package.R
\docType{package}
\name{activityCounts-package}
\alias{activityCounts}
\alias{activityCounts-package}
\title{activityCounts: Generating ActiLife Counts}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Creates ActiLife counts based on the raw acceleration data.
}
\author{
Ruben Brondeel, Javad Rahimipour Anaraki, SeyedJavad KhataeiPour, Daniel Fuller.

\subsection{Maintainers}{SeyedJavad KhataeiPour <skhataeipour@mun.ca>, Javad Rahimipour Anaraki <jra066@mun.ca>, Daniel Fuller <dfuller@mun.ca>}
}
\keyword{internal}
