% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{counts}
\usage{
counts(data, filesf)
}
\arguments{
\item{data}{inout data}

\item{filesf}{sampleing frequency}
}
\value{
returns the number of count which will be used inside tha main function
}
\description{
counts
}
