% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{counts}
\usage{
counts(data, filesf)
}
\arguments{
\item{data}{inout data}

\item{filesf}{sampling frequency}
}
\value{
Returns a \code{data.table} with three columns: \describe{
\item{x}{the number of counts for X axis}
\item{y}{the number of counts for Y axis}
\item{z}{the number of counts for Z axis}
}
}
\description{
counts
}
\examples{
{
\donttest{counts(sampleXYZ,100)}
}


}
