% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{counts}
\usage{
counts(data, hertz = -1, x_axis = 2, y_axis = 3, z_axis = 4,
  time_column = -1, start_time = -1)
}
\arguments{
\item{data}{input accelerometer data}

\item{hertz}{sampling frequency, Hz}

\item{x_axis}{indicates the column number which has the data for x direction, the default is 2}

\item{y_axis}{indicates the column number which has the data for y direction, the default is 3}

\item{z_axis}{indicates the column number which has the data for z direction, the default is 4}

\item{time_column}{indicates the column number which has the date and time}

\item{start_time}{the start date of the measurement}
}
\value{
Returns a \code{data.table} with four columns: \describe{
\item{Time}{The start time of the measurement}
\item{x}{the number of counts for X axis}
\item{y}{the number of counts for Y axis}
\item{z}{the number of counts for Z axis}
}
}
\description{
Calculates ActiLife counts based on raw accelerometer data
}
\examples{
{
\donttest{ ## test with the sampleXYZ dataset whitout start_time
counts(sampleXYZ, hertz = 100,x_axis = 1,y_axis = 2,z_axis = 3)}
\donttest{ ## start time is given
my_start_time <- "2017-08-22 12:30:10"
my_counts <- counts(data = sampleXYZ, hertz = 100,
start_time = my_start_time,x_axis = 1,y_axis = 2,z_axis = 3)}
}

}
\seealso{
\link{sampleXYZ}
 raw accelerometer data for testing \code{counts()} function.

 \link{sampleCounts}
 counts calculated by a\code{ctivityCounts} and ActiLife
}
