% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{counts}
\usage{
counts(data, hertz = -1, x_axis = 1, y_axis = 2, z_axis = 3)
}
\arguments{
\item{data}{input accelerometer data, x y z}

\item{hertz}{sampling frequency, Hz}

\item{x_axis}{indicates the column number which has the data for x direction, the default is 1}

\item{y_axis}{indicates the column number which has the data for y direction, the default is 2}

\item{z_axis}{indicates the column number which has the data for z direction, the default is 3}
}
\value{
Returns a \code{data.table} with three columns: \describe{
\item{x}{the number of counts for X axis}
\item{y}{the number of counts for Y axis}
\item{z}{the number of counts for Z axis}
}
}
\description{
Calculates ActiLife counts based on raw accelerometer data
}
\examples{
{
\donttest{counts(sampleXYZ,100)}
}

}
\seealso{
\link{sampleXYZ}
 raw accelerometer data for testing \code{counts()} function.

 \link{sampleCounts}
 counts calculated by a\code{ctivityCounts} and ActiLife
}
